Start of the rework to deploy from git.
In order to deploy from git, we will need to create a secret token for
azure to use (maybe that is not needed for public repos)
and then we will have to make code changes to the ansible azure
functions.
The python api is autogenerated and maybe we could autogenerate the
ansible wrappers.
I found a guy who wrapped the cli in ansible.